+++ src/Redux/missions/missionReducer.js
@@ -1,5 +1,5 @@
-import getData from './fetchAPI';
-
+// import getData from './fetchAPI';
+// import { useDispatch } from "react-redux";
 export const ADD_STORE = 'ADD_STORE';
 
 export const addStore = (payload) => ({
@@ -7,18 +7,19 @@ export const addStore = (payload) => ({
   payload,
 });
 
-export const displayMissions = () => async (dispatch) => {
-  const missions = await getData();
-  const arr = [];
-  missions.forEach((obj) => {
-    arr.push({
-      mission_id: obj.mission_id,
-      mission_name: obj.mission_name,
-      description: obj.description,
-    });
-  });
-  dispatch(addStore(arr));
-};
+// export const displayMissions = () =>
+// async (dispatch) => {
+//   const missions = await getData();
+//   const arr = [];
+//   missions.forEach((obj) => {
+//     arr.push({
+//       mission_id: obj.mission_id,
+//       mission_name: obj.mission_name,
+//       description: obj.description,
+//     });
+//   });
+//   dispatch(addStore(arr));
+// };
 
 const missionReducer = (state = [], action) => {
   switch (action.type) {
